Trips are now iterative
Resolves all comments
Trips are now iterative, given an array of trips, each trip is populated into current & previous sections
Images are now shown using uri's
