+++ app/screens/Dashboard.tsx
@@ -1,6 +1,4 @@
-import { View, Text } from "react-native"
 import React, { FC } from "react"
-import TabNavigator from "../components/TabNavigator"
 import { Screen } from "../components"
 import { AppStackScreenProps } from "../navigators"
 import { observer } from "mobx-react-lite"
